requirements.txt
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.3.2
idna==3.10
numpy==2.1.2
opencv-python==4.10.0.84
openfoodfacts==1.1.3
pillow==10.4.0
pydantic==2.9.2
pydantic_core==2.23.4
pyzbar==0.1.9
requests==2.32.3
tqdm==4.66.5
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.2.3
